Fix fcitx/fcitx5#56
In current implementation, we only create input context when focus in
happens. But when this happens, it is possible input method is not yet
enabled. While is it possible to create ic on enable the input method,
it would be complex to maintain the state there. Try to always create
ic data on focus in for rest of the case.